requirements.txt
Flask==2.3.2
bluepy==1.3.0
pycryptodome==3.14.1
